Return bool from shared predicates
The FileClass chain (Create, Delete, both Opens, Is_Open,
Do_Is_Available, Is_Available) only ever answered true or false through
an `int`; it now returns bool in the shared file classes and in both
games' CCFileClass and TD's RawFileClass and CDFileClass, and Error()
takes `bool canretry`. Keyboard::Down, the viewport Blit wrappers and
the bignum bit test return bool too, with the SDL state tests and bit
tests written as explicit comparisons.

The VQA movie I/O in both games checked `Open(...) == -1`, which a
true/false result never matched, so a failed open went unnoticed; it
now tests the result directly.

Also: SDL audio calls pass their int flags as 0/1, the bignum carry and
borrow enter arithmetic through explicit casts, and the blitters' int
status codes return 1 rather than true.

+++ src/tech/2keyfbuf.cc
@@ -442,7 +442,8 @@ extern "C" int32_t Buffer_Frame_To_Page(int x, int y, int w, int h, void* src,
   } else {
     // super jump table fun!
     printf("%s new f %x all flags %i\n", __func__,
-           header_pointer->draw_flags & BLIT_ALL, use_all_flags);
+           header_pointer->draw_flags & BLIT_ALL,
+           static_cast<int>(use_all_flags));
   }
 
   return 0;
